HW 04 some work
