+++ HW/Lesson_02/common/utils.py
@@ -10,8 +10,8 @@ from decos import log
 @log
 def get_message(client):
     """
-    Утилита приёма и декодирования сообщения принимает байты выдаёт словарь,
-    если приняточто-то другое отдаёт ошибку значения
+    Утилита приёма и декодирования сообщения принимает байты - выдаёт словарь,
+    если принято что-то другое отдаёт ошибку значения.
     :param client:
     :return:
     """
