+++ HW/Lesson_04/client.py
@@ -452,11 +452,11 @@ def database_load(sock, database, username):
 
 @log
 def main():
-    # Загружаем параметы коммандной строки
+    # Загружаем параметры командной строки
     server_address, server_port, client_name = arg_parser()
 
     """Сообщаем о запуске"""
-    print(f'Консольный месседжер. Клиентский модуль. Имя пользователя: {client_name}')
+    print(f'Консольный мессенджер. Клиентский модуль. Имя пользователя: {client_name}')
 
     # Если имя пользователя не было задано, необходимо запросить пользователя.
     if not client_name:
