+++ HW/Lesson_04/server.ini
@@ -3,3 +3,4 @@ database_path =
 database_file = server_base.db3
 default_port = 7777
 listen_address = 
+
