+++ HW/Lesson_04/server_database.py
@@ -1,165 +1,245 @@
-from pprint import pprint
+"""
+ORM с помощью SQLAalchemy.
 
-from sqlalchemy import create_engine, Table, Column, Integer, String, MetaData, ForeignKey, DateTime
-from sqlalchemy.orm import mapper, sessionmaker
-from common.variables import *
+SQLAalchemy требует предварительной установки:
+у меня необходимо было ещё обновить утилиты
+sudo apt-get install python3-distutils
+pip install sqlalchemy
+
+Декларативный стиль
+"""
+
+from sqlalchemy import create_engine, Column, Integer, String, DateTime, ForeignKey
+from sqlalchemy.ext.declarative import declarative_base
+from sqlalchemy.orm import sessionmaker
 import datetime
 
 
-# Класс - серверная база данных:
+# Класс - серверная база данных
 class ServerDB:
-    # Класс - отображение таблицы всех пользователей
-    class AllUsers:
-        def __init__(self, username):
-            self.name = username
-            self.last_login = datetime.datetime.now()
-            self.id = None
+    Base = declarative_base()
 
-    # Класс - отображение таблицы активных пользователей:
-    class ActiveUsers:
-        def __init__(self, user_id, ip_address, port, login_time):
-            self.user = user_id
-            self.ip_address = ip_address
-            self.port = port
-            self.login_time = login_time
-            self.id = None
+    class AllUsers(Base):
+        """
+        Класс - отображение таблицы всех пользователей
+        """
+        # Создаём таблицу пользователей
+        __tablename__ = 'all_users'
+        id = Column(Integer, primary_key=True)
+        login = Column(String, unique=True)
+        last_conn = Column(DateTime)
 
-    # Класс - отображение таблицы истории входов
-    class LoginHistory:
-        def __init__(self, name, date, ip, port):
-            self.id = None
-            self.name = name
-            self.date_time = date
+        # Экземпляр этого класса - запись в таблице AllUsers
+        def __init__(self, login):
+            self.login = login
+            self.last_conn = datetime.datetime.now()
+
+    class ActiveUsers(Base):
+        """
+        Класс - отображение таблицы активных пользователей:
+        """
+        # Создаём таблицу активностей пользователей
+        __tablename__ = 'active_users'
+        id = Column(Integer, primary_key=True)
+        user = Column(String, ForeignKey('all_users.id'), unique=True)
+        ip = Column(String)
+        port = Column(Integer)
+        time_conn = Column(DateTime)
+
+        #  Экземпляр этого класса - запись в таблице ActiveUsers
+        def __init__(self, user, ip, port, time_conn):
+            self.user = user
             self.ip = ip
             self.port = port
+            self.time_conn = time_conn
 
-    # Класс - отображение таблицы контактов пользователей
-    class UsersContacts:
+    class LoginHistory(Base):
+        """
+        Класс - отображение таблицы истории входов
+        """
+        # Создаём таблицу истории входов в приложение
+        __tablename__ = 'login_history'
+        id = Column(Integer, primary_key=True)
+        user = Column(String, ForeignKey('all_users.id'))
+        ip = Column(String)
+        port = Column(Integer)
+        last_conn = Column(DateTime)
+
+        # Экземпляр этого класса - запись в таблице LoginHistory
+        def __init__(self, user, ip, port, last_conn):
+            self.user = user
+            self.ip = ip
+            self.port = port
+            self.last_conn = last_conn
+
+    class UsersContacts(Base):
+        """
+        Класс - отображение таблицы контактов пользователей
+        """
+        # Создаём таблицу контактов пользователей
+        __tablename__ = 'contacts'
+        id = Column(Integer, primary_key=True)
+        user = Column(String, ForeignKey('all_users.id'))
+        contact = Column(String, ForeignKey('all_users.id'))
+
+        # Экземпляр класса - запись в таблице UsersContacts
         def __init__(self, user, contact):
-            self.id = None
             self.user = user
             self.contact = contact
 
-    # Класс отображение таблицы истории действий
-    class UsersHistory:
+    class UsersHistory(Base):
+        """
+        Класс отображение таблицы истории действий
+        """
+        # Создаём таблицу истории пользователей
+        __tablename__ = 'history'
+        id = Column(Integer, primary_key=True)
+        user = Column(String, ForeignKey('all_users.id'))
+        sent = Column(Integer)
+        accepted = Column(Integer)
+
+        # Экземпляр класса - запись в таблице UsersHistory
         def __init__(self, user):
-            self.id = None
             self.user = user
             self.sent = 0
             self.accepted = 0
 
-    def __init__(self, path):
-        # Создаём движок базы данных
-        print(path)
-        self.database_engine = create_engine(f'sqlite:///{path}', echo=False, pool_recycle=7200,
-                                             connect_args={'check_same_thread': False})
+    def __init__(self):
+        """
+        Создаём движок базы данных
+        SERVER_DATABASE - sqlite:///server_base_pre.db3
+        echo=False - отключает вывод на экран sql-запросов)
+        pool_recycle - по умолчанию соединение с БД через 8 часов простоя обрывается
+        Чтобы этого не случилось необходимо добавить pool_recycle=7200 (переустановка
+        соединения через каждые 2 часа)
+        """
+        self.engine = create_engine('sqlite:///server_base.db3', echo=True, pool_recycle=7200,
+                                    connect_args={'check_same_thread': False})
 
-        # Создаём объект MetaData
-        self.metadata = MetaData()
-
-        # Создаём таблицу пользователей
-        users_table = Table('Users', self.metadata,
-                            Column('id', Integer, primary_key=True),
-                            Column('name', String, unique=True),
-                            Column('last_login', DateTime)
-                            )
-
-        # Создаём таблицу активных пользователей
-        active_users_table = Table('Active_users', self.metadata,
-                                   Column('id', Integer, primary_key=True),
-                                   Column('user', ForeignKey('Users.id'), unique=True),
-                                   Column('ip_address', String),
-                                   Column('port', Integer),
-                                   Column('login_time', DateTime)
-                                   )
-
-        # Создаём таблицу истории входов
-        user_login_history = Table('Login_history', self.metadata,
-                                   Column('id', Integer, primary_key=True),
-                                   Column('name', ForeignKey('Users.id')),
-                                   Column('date_time', DateTime),
-                                   Column('ip', String),
-                                   Column('port', String)
-                                   )
-
-        # Создаём таблицу контактов пользователей
-        contacts = Table('Contacts', self.metadata,
-                         Column('id', Integer, primary_key=True),
-                         Column('user', ForeignKey('Users.id')),
-                         Column('contact', ForeignKey('Users.id'))
-                         )
-
-        # Создаём таблицу истории пользователей
-        users_history_table = Table('History', self.metadata,
-                                    Column('id', Integer, primary_key=True),
-                                    Column('user', ForeignKey('Users.id')),
-                                    Column('sent', Integer),
-                                    Column('accepted', Integer)
-                                    )
-
-        # Создаём таблицы
-        self.metadata.create_all(self.database_engine)
-
-        # Создаём отображения
-        mapper(self.AllUsers, users_table)
-        mapper(self.ActiveUsers, active_users_table)
-        mapper(self.LoginHistory, user_login_history)
-        mapper(self.UsersContacts, contacts)
-        mapper(self.UsersHistory, users_history_table)
-
+        self.Base.metadata.create_all(self.engine)
         # Создаём сессию
-        Session = sessionmaker(bind=self.database_engine)
+        Session = sessionmaker(bind=self.engine)
         self.session = Session()
 
         # Если в таблице активных пользователей есть записи, то их необходимо удалить
+        # Когда устанавливаем соединение, очищаем таблицу активных пользователей
         self.session.query(self.ActiveUsers).delete()
         self.session.commit()
 
-    # Функция выполняющаяся при входе пользователя, записывает в базу факт входа
     def user_login(self, username, ip_address, port):
+        """
+        Функция выполняется при входе пользователя, фиксирует в базе сам факт входа
+        :param username:
+        :param ip_address:
+        :param port:
+        :return:
+        """
         # Запрос в таблицу пользователей на наличие там пользователя с таким именем
-        rez = self.session.query(self.AllUsers).filter_by(name=username)
-
+        rez = self.session.query(self.AllUsers).filter_by(login=username)
+        # print(type(rez))
         # Если имя пользователя уже присутствует в таблице, обновляем время последнего входа
         if rez.count():
             user = rez.first()
-            user.last_login = datetime.datetime.now()
+            user.last_conn = datetime.datetime.now()
         # Если нет, то создаём нового пользователя
         else:
+            # Создаем экземпляр класса self.AllUsers, через который передаем данные в таблицу
             user = self.AllUsers(username)
             self.session.add(user)
-            # Коммит здесь нужен, чтобы присвоился ID
+            # Коммит здесь нужен, чтобы в db записался ID
             self.session.commit()
             user_in_history = self.UsersHistory(user.id)
             self.session.add(user_in_history)
 
         # Теперь можно создать запись в таблицу активных пользователей о факте входа.
+        # Создаем экземпляр класса self.ActiveUsers, через который передаем данные в таблицу
         new_active_user = self.ActiveUsers(user.id, ip_address, port, datetime.datetime.now())
         self.session.add(new_active_user)
 
         # и сохранить в историю входов
-        history = self.LoginHistory(user.id, datetime.datetime.now(), ip_address, port)
+        # Создаем экземпляр класса self.LoginHistory, через который передаем данные в таблицу
+        history = self.LoginHistory(user.id, ip_address, port, datetime.datetime.now())
         self.session.add(history)
 
         # Сохраняем изменения
         self.session.commit()
 
-    # Функция фиксирующая отключение пользователя
     def user_logout(self, username):
+        """
+        Функция фиксирует отключение пользователя
+        и удаляет его из активных пользователей
+        :param username:
+        :return:
+        """
         # Запрашиваем пользователя, что покидает нас
-        user = self.session.query(self.AllUsers).filter_by(name=username).first()
+        # получаем запись из таблицы AllUsers
+        user = self.session.query(self.AllUsers).filter_by(login=username).first()
 
         # Удаляем его из таблицы активных пользователей.
+        # Удаляем запись из таблицы ActiveUsers
         self.session.query(self.ActiveUsers).filter_by(user=user.id).delete()
 
         # Применяем изменения
         self.session.commit()
 
-    # Функция фиксирует передачу сообщения и делает соответствующие отметки в БД
+    def users_list(self):
+        """
+        Функция возвращает список известных пользователей со временем последнего входа.
+        :return:
+        """
+        # Запрос строк таблицы пользователей.
+        query = self.session.query(
+            self.AllUsers.login,
+            self.AllUsers.last_conn,
+        )
+        # Возвращаем список кортежей
+        return query.all()
+
+    def active_users_list(self):
+        """
+        Функция возвращает список активных пользователей
+        :return:
+        """
+        # Запрашиваем соединение таблиц и собираем кортежи - имя, адрес, порт, время.
+        query = self.session.query(
+            self.AllUsers.login,
+            self.ActiveUsers.ip,
+            self.ActiveUsers.port,
+            self.ActiveUsers.time_conn
+        ).join(self.AllUsers)
+
+        # Возвращаем список кортежей
+        return query.all()
+
+    def login_history(self, username=None):
+        """
+        Функция возвращает историю входов по пользователю или по всем пользователям
+        :param username:
+        :return:
+        """
+        # Запрашиваем историю входа
+        query = self.session.query(self.AllUsers.login,
+                                   self.LoginHistory.last_conn,
+                                   self.LoginHistory.ip,
+                                   self.LoginHistory.port
+                                   ).join(self.AllUsers)
+
+        # Если было указано имя пользователя, то фильтруем по нему
+        if username:
+            query = query.filter(self.AllUsers.login == username)
+        return query.all()
+
     def process_message(self, sender, recipient):
+        """
+        Функция фиксирует передачу сообщения и делает соответствующие отметки в БД
+        :param sender:
+        :param recipient:
+        :return:
+        """
         # Получаем ID отправителя и получателя
-        sender = self.session.query(self.AllUsers).filter_by(name=sender).first().id
-        recipient = self.session.query(self.AllUsers).filter_by(name=recipient).first().id
+        sender = self.session.query(self.AllUsers).filter_by(login=sender).first().id
+        recipient = self.session.query(self.AllUsers).filter_by(login=recipient).first().id
+
         # Запрашиваем строки из истории и увеличиваем счётчики
         sender_row = self.session.query(self.UsersHistory).filter_by(user=sender).first()
         sender_row.sent += 1
@@ -168,11 +248,16 @@ class ServerDB:
 
         self.session.commit()
 
-    # Функция добавляет контакт для пользователя.
     def add_contact(self, user, contact):
+        """
+        Функция добавляет контакт для пользователя.
+        :param user:
+        :param contact:
+        :return:
+        """
         # Получаем ID пользователей
-        user = self.session.query(self.AllUsers).filter_by(name=user).first()
-        contact = self.session.query(self.AllUsers).filter_by(name=contact).first()
+        user = self.session.query(self.AllUsers).filter_by(login=user).first()
+        contact = self.session.query(self.AllUsers).filter_by(login=contact).first()
 
         # Проверяем что не дубль и что контакт может существовать (полю пользователь мы доверяем)
         if not contact or self.session.query(self.UsersContacts).filter_by(user=user.id, contact=contact.id).count():
@@ -183,11 +268,16 @@ class ServerDB:
         self.session.add(contact_row)
         self.session.commit()
 
-    # Функция удаляет контакт из базы данных
     def remove_contact(self, user, contact):
+        """
+        Функция удаляет контакт из базы данных
+        :param user:
+        :param contact:
+        :return:
+        """
         # Получаем ID пользователей
-        user = self.session.query(self.AllUsers).filter_by(name=user).first()
-        contact = self.session.query(self.AllUsers).filter_by(name=contact).first()
+        user = self.session.query(self.AllUsers).filter_by(login=user).first()
+        contact = self.session.query(self.AllUsers).filter_by(login=contact).first()
 
         # Проверяем что контакт может существовать (полю пользователь мы доверяем)
         if not contact:
@@ -200,79 +290,62 @@ class ServerDB:
         ).delete())
         self.session.commit()
 
-    # Функция возвращает список известных пользователей со временем последнего входа.
-    def users_list(self):
-        # Запрос строк таблицы пользователей.
-        query = self.session.query(
-            self.AllUsers.name,
-            self.AllUsers.last_login
-        )
-        # Возвращаем список кортежей
-        return query.all()
-
-    # Функция возвращает список активных пользователей
-    def active_users_list(self):
-        # Запрашиваем соединение таблиц и собираем кортежи имя, адрес, порт, время.
-        query = self.session.query(
-            self.AllUsers.name,
-            self.ActiveUsers.ip_address,
-            self.ActiveUsers.port,
-            self.ActiveUsers.login_time
-        ).join(self.AllUsers)
-        # Возвращаем список кортежей
-        return query.all()
-
-    # Функция возвращающая историю входов по пользователю или всем пользователям
-    def login_history(self, username=None):
-        # Запрашиваем историю входа
-        query = self.session.query(self.AllUsers.name,
-                                   self.LoginHistory.date_time,
-                                   self.LoginHistory.ip,
-                                   self.LoginHistory.port
-                                   ).join(self.AllUsers)
-        # Если было указано имя пользователя, то фильтруем по нему
-        if username:
-            query = query.filter(self.AllUsers.name == username)
-        # Возвращаем список кортежей
-        return query.all()
-
-    # Функция возвращает список контактов пользователя.
     def get_contacts(self, username):
+        """
+        Функция возвращает список контактов пользователя.
+        :param username:
+        :return:
+        """
         # Запрашиваем указанного пользователя
-        user = self.session.query(self.AllUsers).filter_by(name=username).one()
+        user = self.session.query(self.AllUsers).filter_by(login=username).one()
 
         # Запрашиваем его список контактов
         query = self.session.query(self.UsersContacts, self.AllUsers.name). \
-            filter_by(user=user.id). \
+            filter_by(login=user.id). \
             join(self.AllUsers, self.UsersContacts.contact == self.AllUsers.id)
 
         # выбираем только имена пользователей и возвращаем их.
         return [contact[1] for contact in query.all()]
 
-    # Функция возвращает количество переданных и полученных сообщений
     def message_history(self):
+        """
+        Функция возвращает количество переданных и полученных сообщений.
+        :return:
+        """
         query = self.session.query(
-            self.AllUsers.name,
-            self.AllUsers.last_login,
+            self.AllUsers.login,
+            self.AllUsers.last_conn,
             self.UsersHistory.sent,
             self.UsersHistory.accepted
         ).join(self.AllUsers)
+
         # Возвращаем список кортежей
         return query.all()
 
 
 # Отладка
 if __name__ == '__main__':
-    test_db = ServerStorage('server_base_pre.db3')
-    test_db.user_login('1111', '192.168.1.113', 8080)
-    test_db.user_login('McG2', '192.168.1.113', 8081)
-    pprint(test_db.users_list())
-    pprint(test_db.active_users_list())
-    test_db.user_logout('McG2')
-    pprint(test_db.login_history('re'))
-    test_db.add_contact('test2', 'test1')
-    test_db.add_contact('test1', 'test3')
-    test_db.add_contact('test1', 'test6')
-    test_db.remove_contact('test1', 'test3')
-    test_db.process_message('McG2', '1111')
-    pprint(test_db.message_history())
+    db = ServerDB()
+
+    # Выполняем "подключение" пользователя
+    db.user_login('client_1', '192.168.1.4', 7778)
+    db.user_login('client_2', '192.168.1.5', 7777)
+
+    # выводим список кортежей - активных пользователей
+    print(db.active_users_list())
+
+    # выполняем 'отключение' пользователя
+    db.user_logout('client_1')
+    print(db.users_list())
+
+    # выводим список активных пользователей
+    print(db.active_users_list())
+    db.user_logout('client_2')
+    print(db.users_list())
+    print(db.active_users_list())
+
+    # запрашиваем историю входов по пользователю
+    db.login_history('client_1')
+
+    # выводим список известных пользователей
+    print(db.users_list())
