--- HW/Lesson_04/server_database_decl..py
@@ -1,230 +1,0 @@
-"""
-ORM с помощью SQLAalchemy.
-
-SQLAalchemy требует предварительной установки:
-у меня необходимо было ещё обновить утилиты
-sudo apt-get install python3-distutils
-pip install sqlalchemy
-
-Декларативный стиль
-"""
-
-from sqlalchemy import create_engine, Column, Integer, String, DateTime, ForeignKey
-from sqlalchemy.ext.declarative import declarative_base
-from sqlalchemy.orm import sessionmaker
-import datetime
-
-
-# Класс - серверная база данных
-class ServerDB:
-    Base = declarative_base()
-
-
-    class AllUsers(Base):
-        """
-        Класс - отображение таблицы всех пользователей
-        """
-        # Создаём таблицу пользователей
-        __tablename__ = 'all_users'
-        id = Column(Integer, primary_key=True)
-        login = Column(String, unique=True)
-        last_conn = Column(DateTime)
-
-        # Экземпляр этого класса - запись в таблице AllUsers
-        def __init__(self, login):
-            self.login = login
-            self.last_conn = datetime.datetime.now()
-
-
-
-    class ActiveUsers(Base):
-        """
-        Класс - отображение таблицы активных пользователей:
-        """
-        # Создаём таблицу активностей пользователей
-        __tablename__ = 'active_users'
-        id = Column(Integer, primary_key=True)
-        user = Column(String, ForeignKey('all_users.id'), unique=True)
-        ip = Column(String)
-        port = Column(Integer)
-        time_conn = Column(DateTime)
-
-        #  Экземпляр этого класса - запись в таблице ActiveUsers
-        def __init__(self, user, ip, port, time_conn):
-            self.user = user
-            self.ip = ip
-            self.port = port
-            self.time_conn = time_conn
-
-
-    class LoginHistory(Base):
-        """
-        Класс - отображение таблицы истории входов
-        """
-        # Создаём таблицу истории входов в приложение
-        __tablename__ = 'login_history'
-        id = Column(Integer, primary_key=True)
-        user = Column(String, ForeignKey('all_users.id'))
-        ip = Column(String)
-        port = Column(Integer)
-        last_conn = Column(DateTime)
-
-        # Экземпляр этого класса - запись в таблице LoginHistory
-        def __init__(self, user, ip, port, last_conn):
-            self.user = user
-            self.ip = ip
-            self.port = port
-            self.last_conn = last_conn
-
-    def __init__(self):
-        """
-        Создаём движок базы данных
-        SERVER_DATABASE - sqlite:///server_base_pre.db3
-        echo=False - отключает вывод на экран sql-запросов)
-        pool_recycle - по умолчанию соединение с БД через 8 часов простоя обрывается
-        Чтобы этого не случилось необходимо добавить pool_recycle=7200 (переустановка
-        соединения через каждые 2 часа)
-        """
-        self.engine = create_engine('sqlite:///server_base_pre.db3', echo=True, pool_recycle=7200)
-
-        self.Base.metadata.create_all(self.engine)
-        # Создаём сессию
-        Session = sessionmaker(bind=self.engine)
-        self.session = Session()
-
-        # Если в таблице активных пользователей есть записи, то их необходимо удалить
-        # Когда устанавливаем соединение, очищаем таблицу активных пользователей
-        self.session.query(self.ActiveUsers).delete()
-        self.session.commit()
-
-
-    def user_login(self, username, ip_address, port):
-        """
-        Функция выполняется при входе пользователя, фиксирует в базе сам факт входа
-        :param username:
-        :param ip_address:
-        :param port:
-        :return:
-        """
-        # Запрос в таблицу пользователей на наличие там пользователя с таким именем
-        rez = self.session.query(self.AllUsers).filter_by(login=username)
-        # print(type(rez))
-        # Если имя пользователя уже присутствует в таблице, обновляем время последнего входа
-        if rez.count():
-            user = rez.first()
-            user.last_conn = datetime.datetime.now()
-        # Если нет, то создаём нового пользователя
-        else:
-            # Создаем экземпляр класса self.AllUsers, через который передаем данные в таблицу
-            user = self.AllUsers(username)
-            self.session.add(user)
-            # Коммит здесь нужен, чтобы в db записался ID
-            self.session.commit()
-
-        # Теперь можно создать запись в таблицу активных пользователей о факте входа.
-        # Создаем экземпляр класса self.ActiveUsers, через который передаем данные в таблицу
-        new_active_user = self.ActiveUsers(user.id, ip_address, port, datetime.datetime.now())
-        self.session.add(new_active_user)
-
-        # и сохранить в историю входов
-        # Создаем экземпляр класса self.LoginHistory, через который передаем данные в таблицу
-        history = self.LoginHistory(user.id, ip_address, port, datetime.datetime.now())
-        self.session.add(history)
-
-        # Сохраняем изменения
-        self.session.commit()
-
-
-    def user_logout(self, username):
-        """
-        Функция фиксирует отключение пользователя
-        и удаляет его из активных пользователей
-        :param username:
-        :return:
-        """
-        # Запрашиваем пользователя, что покидает нас
-        # получаем запись из таблицы AllUsers
-        user = self.session.query(self.AllUsers).filter_by(login=username).first()
-
-        # Удаляем его из таблицы активных пользователей.
-        # Удаляем запись из таблицы ActiveUsers
-        self.session.query(self.ActiveUsers).filter_by(user=user.id).delete()
-
-        # Применяем изменения
-        self.session.commit()
-
-
-    def users_list(self):
-        """
-        Функция возвращает список известных пользователей со временем последнего входа.
-        :return:
-        """
-        query = self.session.query(
-            self.AllUsers.login,
-            self.AllUsers.last_conn,
-        )
-        # Возвращаем список тюплов
-        return query.all()
-
-
-    def active_users_list(self):
-        """
-        Функция возвращает список активных пользователей
-        :return:
-        """
-        # Запрашиваем соединение таблиц и собираем тюплы имя, адрес, порт, время.
-        query = self.session.query(
-            self.AllUsers.login,
-            self.ActiveUsers.ip,
-            self.ActiveUsers.port,
-            self.ActiveUsers.time_conn
-        ).join(self.AllUsers)
-        # Возвращаем список тюплов
-        return query.all()
-
-
-    def login_history(self, username=None):
-        """
-        Функция возвращает историю входов по пользователю или по всем пользователям
-        :param username:
-        :return:
-        """
-        # Запрашиваем историю входа
-        query = self.session.query(self.AllUsers.login,
-                                   self.LoginHistory.last_conn,
-                                   self.LoginHistory.ip,
-                                   self.LoginHistory.port
-                                   ).join(self.AllUsers)
-        # Если было указано имя пользователя, то фильтруем по нему
-        if username:
-            query = query.filter(self.AllUsers.login == username)
-        return query.all()
-
-
-# Отладка
-if __name__ == '__main__':
-
-    db = ServerDB()
-
-    # Выполняем "подключение" пользователя
-    db.user_login('client_1', '192.168.1.4', 7778)
-    db.user_login('client_2', '192.168.1.5', 7777)
-
-    # выводим список кортежей - активных пользователей
-    print(db.active_users_list())
-
-    # выполняем 'отключение' пользователя
-    db.user_logout('client_1')
-    print(db.users_list())
-
-    # выводим список активных пользователей
-    print(db.active_users_list())
-    db.user_logout('client_2')
-    print(db.users_list())
-    print(db.active_users_list())
-
-    # запрашиваем историю входов по пользователю
-    db.login_history('client_1')
-
-    # выводим список известных пользователей
-    print(db.users_list())
